part 4 complete

+++ controllers/blogController.js
@@ -1,22 +1,41 @@
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
 
 blogRouter.get("/", async (request, response) => {
-  const notes = await Blog.find({});
-  response.json(notes.map((note) => note.toJSON()));
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+  response.json(blogs.map((blog) => blog.toJSON()));
 });
 
 blogRouter.post("/", async (request, response, next) => {
+  const body = request.body;
+  const token = request.token;
   try {
+    if (!token) {
+      return response.status(401).json({ error: "Token missing" });
+    }
+
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: "Token Invalid" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+
     const blog = new Blog({
       author: request.body.author,
       likes: request.body.likes || 0,
       title: request.body.title,
       url: request.body.url,
+      user: user._id,
     });
 
-    const saveBlog = await blog.save();
-    response.json(saveBlog.toJSON());
+    const savedNote = await blog.save();
+    user.blogs = user.blogs.concat(savedNote._id);
+    await user.save();
+    response.json(savedNote.toJSON());
   } catch (exception) {
     next(exception);
   }
@@ -24,9 +43,21 @@ blogRouter.post("/", async (request, response, next) => {
 
 blogRouter.delete("/:id", async (request, response, next) => {
   const id = request.params.id;
+
+  const token = request.token;
+  console.log(token);
   try {
-    await Blog.findByIdAndRemove(id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(token, process.env.SECRET);
+    const blog = await Blog.findById(id);
+
+    if (decodedToken.id.toString() === blog.user.toString()) {
+      await Blog.findByIdAndRemove(id);
+      response.status(204).end();
+    } else {
+      response
+        .status(401)
+        .send({ error: "You have no permission to delete tha blog" });
+    }
   } catch (error) {
     next(error);
   }
